server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
pyotp==2.9.0
qrcode==7.4.2
argon2-cffi==23.1.0
cryptography==41.0.7
Pillow==10.2.0
